requirements-airflow.txt
boto3
pandas
pyarrow
python-dotenv
requests
sqlalchemy
psycopg2-binary
xgboost
lightgbm
prophet
optuna
mlflow
scikit-learn